# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FluxPay",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/console/pay": {
      "post": {
        "tags": [
          "unified_console"
        ],
        "summary": "Execute Payment",
        "description": "Execute live payment between agent and merchant or recipient agent (A2A).",
        "operationId": "execute_payment_api_console_pay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Execute Payment Api Console Pay Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/console/a2a": {
      "post": {
        "tags": [
          "unified_console"
        ],
        "summary": "Execute A2A Payment",
        "description": "Execute live autonomous Agent-to-Agent (A2A / M2M) machine transfer.",
        "operationId": "execute_a2a_payment_api_console_a2a_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/A2ARequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Execute A2A Payment Api Console A2A Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/console/agents": {
      "post": {
        "tags": [
          "unified_console"
        ],
        "summary": "Create Agent",
        "description": "Create a new autonomous agent.",
        "operationId": "create_agent_api_console_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Agent Api Console Agents Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/console/agents/{agent_id}/toggle": {
      "post": {
        "tags": [
          "unified_console"
        ],
        "summary": "Toggle Agent",
        "description": "Toggle agent status (active / suspended).",
        "operationId": "toggle_agent_api_console_agents__agent_id__toggle_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Toggle Agent Api Console Agents  Agent Id  Toggle Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/console/agents/{agent_id}/topup": {
      "post": {
        "tags": [
          "unified_console"
        ],
        "summary": "Topup Agent",
        "description": "Fund an agent account.",
        "operationId": "topup_agent_api_console_agents__agent_id__topup_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopupAgentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Topup Agent Api Console Agents  Agent Id  Topup Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/console/deposit": {
      "post": {
        "tags": [
          "unified_console"
        ],
        "summary": "Deposit Funds",
        "description": "Confirm incoming crypto on-ramp deposit for an agent with real on-chain validation.",
        "operationId": "deposit_funds_api_console_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Deposit Funds Api Console Deposit Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/console/inspect-tx": {
      "post": {
        "tags": [
          "unified_console"
        ],
        "summary": "Inspect Blockchain Tx",
        "description": "Inspect and decode on-chain transaction amount from Solana or TRON.",
        "operationId": "inspect_blockchain_tx_api_console_inspect_tx_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectTxRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Inspect Blockchain Tx Api Console Inspect Tx Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/console/merchants": {
      "post": {
        "tags": [
          "unified_console"
        ],
        "summary": "Create Merchant",
        "description": "Register a new merchant.",
        "operationId": "create_merchant_api_console_merchants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMerchantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Merchant Api Console Merchants Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/console/holds/{hold_id}/action": {
      "post": {
        "tags": [
          "unified_console"
        ],
        "summary": "Hold Action",
        "description": "Approve or reject a quarantined payment.",
        "operationId": "hold_action_api_console_holds__hold_id__action_post",
        "parameters": [
          {
            "name": "hold_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Hold Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Hold Action Api Console Holds  Hold Id  Action Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/console/limits": {
      "post": {
        "tags": [
          "unified_console"
        ],
        "summary": "Update Limits",
        "description": "Update risk limits for an agent.",
        "operationId": "update_limits_api_console_limits_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLimitsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Update Limits Api Console Limits Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "A2ARequest": {
        "properties": {
          "from_agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "From Agent Id",
            "description": "Paying Agent UUID"
          },
          "to_agent": {
            "type": "string",
            "title": "To Agent",
            "description": "Recipient Agent UUID or external_id"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "description": "Amount in USDC"
          }
        },
        "type": "object",
        "required": [
          "from_agent_id",
          "to_agent",
          "amount"
        ],
        "title": "A2ARequest"
      },
...
      "CreateAgentRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          },
          "external_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "External Id"
          },
          "initial_balance": {
            "type": "number",
            "minimum": 0.0,
            "title": "Initial Balance",
            "default": 1000.0
          }
        },
        "type": "object",
        "required": [
          "name",
          "external_id"
        ],
        "title": "CreateAgentRequest"
      },
      "CreateMerchantRequest": {
        "properties": {
          "external_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "External Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          },
          "webhook_url": {
            "type": "string",
            "title": "Webhook Url",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "external_id",
          "name"
        ],
        "title": "CreateMerchantRequest"
      },
      "DepositRequest": {
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "network": {
            "type": "string",
            "title": "Network",
            "default": "SOLANA"
          },
          "tx_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Hash"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "amount"
        ],
        "title": "DepositRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HoldActionRequest": {
        "properties": {
          "action": {
            "type": "string",
            "pattern": "^(approve|reject)$",
            "title": "Action"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "HoldActionRequest"
      },
      "InspectTxRequest": {
        "properties": {
          "network": {
            "type": "string",
            "title": "Network",
            "default": "SOLANA"
          },
          "tx_hash": {
            "type": "string",
            "title": "Tx Hash"
          }
        },
        "type": "object",
        "required": [
          "tx_hash"
        ],
        "title": "InspectTxRequest"
      },
      "PayRequest": {
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "to_merchant": {
            "type": "string",
            "title": "To Merchant"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "description": "Amount in USDC"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "to_merchant",
          "amount"
        ],
        "title": "PayRequest"
      },
...
      "TopupAgentRequest": {
        "properties": {
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "TopupAgentRequest"
      },
      "UpdateLimitsRequest": {
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "single_max": {
            "type": "number",
            "minimum": 1.0,
            "title": "Single Max"
          },
          "daily_max": {
            "type": "number",
            "minimum": 1.0,
            "title": "Daily Max"
          },
          "velocity_count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Velocity Count",
            "default": 10
          },
          "velocity_window_s": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Velocity Window S",
            "default": 60
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "single_max",
          "daily_max"
        ],
        "title": "UpdateLimitsRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}